ported changes from ZDoom revision 3447-3450 and from GZDoom revision 1315-1316
* Updated to ZDoom r3450:
- Fix signed/unsigned mismatch warned by GCC.
- Moved "Go away!" text into language.enu.
- In conjunction with all the below changes, attempt to fix A_CheckSightOrRange and A_CheckSight for multiplayer: They now always check through the eyes of every player. For players whose cameras are not players, they also check through the eyes of those cameras.
- Using spynext/spyprev to switch from a non-player to a player now writes a command to the network stream and lets Net_DoCommand() take care of it later. The logic here is that if a player is viewing from something that isn't another player, then every player needs to know about it for sync purposes. Consequently, when they stop viewing from a non-player and switch to a player, everybody needs to know about that too. But if they are viewing from a player, it doesn't matter which player it is, so they can spynext/spyprev all they want without letting the other players know about it (and without potentially breaking demos--due to the above-mentioned two codepointers--while doing it during demo playback).
- Replaced the instances of checking players[consoleplayer].camera for a valid pointer to ones that do it for every player.
- Fixed: Upon changing levels, all players but the consoleplayer would have their cameras NULLed.
- Fixed: player_t::FixPointers() needs to bypass the read barriers, or it won't be able to do substitutions of old objects that are pending deletion.
- Revised the fix from r3442: Make the line a nonrepeatable Door_Open instead of completely clearing the line's special.

git-svn-id: http://mancubus.net/svn/hosted/gzdoom/trunk@1315 b0f79afe-0144-0410-b225-9a4edf0717df

- Fixed what looks like a copy/paste mistake.

git-svn-id: http://mancubus.net/svn/hosted/gzdoom/trunk@1316 b0f79afe-0144-0410-b225-9a4edf0717df

+++ src/p_acs.cpp
@@ -7230,6 +7230,9 @@ int DLevelScript::RunScript ()
 			STACK(1) = !STACK(1);
 			break;
 
+
+
+
 		case PCD_NEGATEBINARY:
 			STACK(1) = ~STACK(1);
 			break;
@@ -9721,11 +9724,11 @@ static void addDefered (level_info_t *i, acsdefered_t::EType type, int script, c
 		def->next = i->defered;
 		def->type = type;
 		def->script = script;
-		for (j = 0; j < countof(def->args) && j < argcount; ++j)
+		for (j = 0; (size_t)j < countof(def->args) && j < argcount; ++j)
 		{
 			def->args[j] = args[j];
 		}
-		while (j < countof(def->args))
+		while ((size_t)j < countof(def->args))
 		{
 			def->args[j++] = 0;
 		}
